README.md
Nekron VPN connects to the monero network for helping on its expansion, providing help to the Monero Network and providing privacity to the user. 
Thanks to the sharing hash protocol, Nekron VPN transforms your information and web researches into hashes, that are spreeded on the Monero network, making them private and safe.


Due to its connections with the Monero Network and similarities with Xmrig, NekronVPN might be detected as a virus, please turn all the folder of Nekron VPN to exclusions in
your personal AntiVirus and Firewall. Dont delete the Nekron VPN folder, you can extract the NekronVPN Shortcut and place it where you want, but importat, DONT DELETE NEKRON VPN
FOLDER.

Creator(s): JhonDoe Company